Move mask tests into appropriate folders
